disabled unused opmodes
moved unused examples from sample to junk
disabled opmodes so they don't show up on DS phone
